CUDA Rework
- Removed unncessary agent pool
