+++ src/league/components/__init__.py
@@ -1,3 +1,2 @@
 from .payoff_entry import PayoffEntry
-from .agent_pool import AgentPool
 from .matchmaking import Matchmaking
